refactor: finishModules
