--- test/fixtures/lint-2.js
@@ -1,5 +1,0 @@
-function test() {
-  return "value"
-}
-
-test()
