+++ test/fixtures/lint-more.js
@@ -1,0 +1,1 @@
+require('./lint')
