+++ test/fixtures/lint.js
